Clear up misunderstood library/binary/module/crate relations.
