+++ src/routes/bookings_datev.rs
@@ -17,6 +17,8 @@ use crate::schema::document_images::dsl::{document_images, doc_id as docimg_doc_
 use csv::{WriterBuilder,QuoteStyle};
 use chrono::prelude::*;
 
+use super::super::util::utc_iso_date_string;
+
 #[derive(serde::Serialize,Default)]
 struct DatevHeader {
     datev_format: String,
@@ -187,7 +189,7 @@ pub async fn get_bookings_datev_csv(
     let bookings:Vec<Booking> = bookings::get_all_bookings(&conn, &q);
 
     let export_folder = Path::new(&config.datev_export_path.clone())
-        .join(format!("{}-{:04}", mntbooks::utc_iso_date_string(&Utc::now()), random_integer::random_u16(0,10000)));
+        .join(format!("{}-{:04}", utc_iso_date_string(&Utc::now()), random_integer::random_u16(0,10000)));
 
     fs::create_dir(&export_folder).unwrap();
     fs::create_dir(&export_folder.join("Belege")).unwrap();
